requirements-airflow.txt
apache-airflow >= 2.10.0
apache-airflow-providers-apache-kafka >= 1.10.3rc1
apache-airflow-providers-celery >= 3.3.0
apache-airflow-providers-common-compat >= 1.7.3
apache-airflow-providers-common-io >= 1.6.2
apache-airflow-providers-common-sql >= 1.27.5
apache-airflow-providers-elasticsearch >= 5.3.0
apache-airflow-providers-fab >= 1.5.3
apache-airflow-providers-ftp >= 3.13.2
apache-airflow-providers-http >= 5.3.3
apache-airflow-providers-imap >= 3.9.2
apache-airflow-providers-postgres >= 5.6.0
apache-airflow-providers-redis >= 3.4.1
apache-airflow-providers-smtp >= 2.2.0
apache-airflow-providers-sqlite >= 4.1.2
psycopg2-binary >= 2.9.10
redis >= 4.6.0
SQLAlchemy >= 1.4.54
SQLAlchemy-JSONField >= 1.0.2
SQLAlchemy-Utils >= 0.42.0
PyYAML >= 6.0.2
tqdm >= 4.67.1
traitlets >= 5.14.3
